Rename `purchases_flutter_ui` to `purchases_ui_flutter` (#921)
This renames the `purchases_flutter_ui` package added in #919 to
`purchases_ui_flutter`
