+++ .circleci/config.yml
@@ -167,13 +167,13 @@ jobs:
       - flutter-get-dependencies:
           project: api_tester
       - flutter-get-dependencies:
-          project: purchases_flutter_ui
+          project: purchases_ui_flutter
       - run:
           name: Analyze code
           command: flutter analyze
       - run:
-          name: Analyze code for purchases_flutter_ui
-          command: cd purchases_flutter_ui && flutter analyze
+          name: Analyze code for purchases_ui_flutter
+          command: cd purchases_ui_flutter && flutter analyze
 
   test:
     description: "Run tests for Flutter"
@@ -186,8 +186,8 @@ jobs:
           name: Run tests
           command: flutter test
       - run:
-          name: Run tests for purchases_flutter_ui
-          command: cd purchases_flutter_ui && flutter test
+          name: Run tests for purchases_ui_flutter
+          command: cd purchases_ui_flutter && flutter test
 
   api_test:
     description: "Run API tests for Flutter"
@@ -325,8 +325,8 @@ jobs:
           name: Dry Run Release
           command: flutter pub publish --dry-run
       - run:
-          name: Dry Run Release for purchases_flutter_ui
-          command: cd purchases_flutter_ui && flutter pub publish --dry-run
+          name: Dry Run Release for purchases_ui_flutter
+          command: cd purchases_ui_flutter && flutter pub publish --dry-run
 
   update-hybrid-common-versions:
     description: "Creates a PR updating purchases-hybrid-common to latest release"
@@ -371,7 +371,7 @@ jobs:
       - revenuecat/trust-github-key
       - run:
           name: release
-          command: bundle exec fastlane release_purchases_flutter_ui
+          command: bundle exec fastlane release_purchases_ui_flutter
 
   github_release:
     description: "Creates a github release with the current version"
