+++ api_tester/pubspec.yaml
@@ -44,8 +44,8 @@ dev_dependencies:
 
   purchases_flutter:
     path: ../
-  purchases_flutter_ui:
-    path: ../purchases_flutter_ui
+  purchases_ui_flutter:
+    path: ../purchases_ui_flutter
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
