+++ fastlane/Fastfile
@@ -25,10 +25,10 @@ files_with_version_number = {
     './ios/Classes/PurchasesFlutterPlugin.m' => ['return @"{x}"'],
     './android/build.gradle' => ['version \'{x}\''],
     './android/src/main/java/com/revenuecat/purchases_flutter/PurchasesFlutterPlugin.java' => ['PLUGIN_VERSION = "{x}"'],
-    './purchases_flutter_ui/pubspec.yaml' => ['version: {x}'],
-    './purchases_flutter_ui/ios/purchases_flutter_ui.podspec' => ['s.version          = \'{x}\''],
-    './purchases_flutter_ui/macos/purchases_flutter_ui.podspec' => ['s.version          = \'{x}\''],
-    './purchases_flutter_ui/android/build.gradle' => ['version \'{x}\''],
+    './purchases_ui_flutter/pubspec.yaml' => ['version: {x}'],
+    './purchases_ui_flutter/ios/purchases_ui_flutter.podspec' => ['s.version          = \'{x}\''],
+    './purchases_ui_flutter/macos/purchases_ui_flutter.podspec' => ['s.version          = \'{x}\''],
+    './purchases_ui_flutter/android/build.gradle' => ['version \'{x}\''],
 }
 
 repo_name = 'purchases-flutter'
@@ -89,10 +89,10 @@ end
 
 desc "Create release"
 lane :release do |options|
-  # Remove purchases_flutter_ui so it's not included in the release of the main SDK
+  # Remove purchases_ui_flutter so it's not included in the release of the main SDK
   # Can't use .pubignore because it causes files to not be found when trying to
-  # deploy purchases_flutter_ui
-  sh('rm', '-rf', 'purchases_flutter_ui')
+  # deploy purchases_ui_flutter
+  sh('rm', '-rf', 'purchases_ui_flutter')
   Dir.chdir('..') do
     sh('flutter', 'pub', 'publish', '--dry-run')
   end
@@ -102,13 +102,13 @@ lane :release do |options|
   end
 end
 
-desc "Create purchases_flutter_ui release"
-lane :release_purchases_flutter_ui do |options|
-  Dir.chdir('../purchases_flutter_ui') do
+desc "Create purchases_ui_flutter release"
+lane :release_purchases_ui_flutter do |options|
+  Dir.chdir('../purchases_ui_flutter') do
     sh('flutter', 'pub', 'publish', '--dry-run')
   end
   setup_credentials_file
-  Dir.chdir('../purchases_flutter_ui') do
+  Dir.chdir('../purchases_ui_flutter') do
     sh('flutter', 'pub', 'publish', '--force')
   end
 end
@@ -164,9 +164,9 @@ lane :update_hybrid_common do |options|
       'ios/purchases_flutter.podspec' => ['s.dependency \'PurchasesHybridCommon\', \'{x}\''],
       'macos/purchases_flutter.podspec' => ['s.dependency \'PurchasesHybridCommon\', \'{x}\''],
       'android/build.gradle' => ['ext.common_version = \'{x}\''],
-      'purchases_flutter_ui/ios/purchases_flutter_ui.podspec' => ['s.dependency \'PurchasesHybridCommon\', \'{x}\''],
-      'purchases_flutter_ui/macos/purchases_flutter_ui.podspec' => ['s.dependency \'PurchasesHybridCommon\', \'{x}\''],
-      'purchases_flutter_ui/android/build.gradle' => ['ext.common_version = \'{x}\''],
+      'purchases_ui_flutter/ios/purchases_ui_flutter.podspec' => ['s.dependency \'PurchasesHybridCommon\', \'{x}\''],
+      'purchases_ui_flutter/macos/purchases_ui_flutter.podspec' => ['s.dependency \'PurchasesHybridCommon\', \'{x}\''],
+      'purchases_ui_flutter/android/build.gradle' => ['ext.common_version = \'{x}\''],
   }
 
   if dry_run
