--- purchases_flutter_ui/android/settings.gradle
@@ -1,1 +1,0 @@
-rootProject.name = 'purchases_flutter_ui'
