--- purchases_flutter_ui/macos/Classes/PurchasesFlutterUiPlugin.swift
@@ -1,1 +1,0 @@
-../../ios/Classes/PurchasesFlutterUiPlugin.swift
