+++ purchases_ui_flutter/android/settings.gradle
@@ -1,0 +1,1 @@
+rootProject.name = 'purchases_ui_flutter'
