+++ purchases_ui_flutter/android/src/main/AndroidManifest.xml
@@ -1,3 +1,3 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-  package="com.revenuecat.purchases_flutter_ui">
+  package="com.revenuecat.purchases_ui_flutter">
 </manifest>
