+++ purchases_ui_flutter/example/README.md
@@ -3,4 +3,4 @@ Our examples are located at [revenuecat_examples](https://github.com/RevenueCat/
 
 ### [Purchase Tester](https://github.com/RevenueCat/purchases-flutter/tree/main/revenuecat_examples/purchase_tester)
 
-Contains an example app that showcases how to use `purchases_flutter` and `purchases_flutter_ui` to display a paywall and make purchases.
+Contains an example app that showcases how to use `purchases_flutter` and `purchases_ui_flutter` to display a paywall and make purchases.
