+++ purchases_ui_flutter/macos/Classes/PurchasesUiFlutterPlugin.swift
@@ -1,0 +1,1 @@
+../../ios/Classes/PurchasesUiFlutterPlugin.swift
