+++ purchases_ui_flutter/macos/purchases_ui_flutter.podspec
@@ -1,9 +1,9 @@
 #
 # To learn more about a Podspec see http://guides.cocoapods.org/syntax/podspec.html.
-# Run `pod lib lint purchases_flutter_ui.podspec` to validate before publishing.
+# Run `pod lib lint purchases_ui_flutter.podspec` to validate before publishing.
 #
 Pod::Spec.new do |s|
-  s.name             = 'purchases_flutter_ui'
+  s.name             = 'purchases_ui_flutter'
   s.version          = '6.6.0-beta.3'
   s.summary          = 'Flutter plugin that integrates RevenueCat Paywalls'
   s.description      = <<-DESC
@@ -12,13 +12,13 @@ Flutter plugin that integrates RevenueCat Paywalls
   s.homepage         = 'http://revenuecat.com'
   s.license          = { :file => '../LICENSE' }
   s.author           = { 'RevenueCat' => 'support@revenuecat.com' }
-  s.source           = { :path => '.' }
-  s.source_files = 'Classes/**/*'
-  s.dependency 'Flutter'
-  s.dependency 'PurchasesHybridCommon', '8.10.0-beta.8'
-  s.platform = :ios, '11.0'
 
-  # Flutter.framework does not contain a i386 slice.
-  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES', 'EXCLUDED_ARCHS[sdk=iphonesimulator*]' => 'i386' }
+  s.source           = { :path => '.' }
+  s.source_files     = 'Classes/**/*'
+  s.dependency 'FlutterMacOS'
+  s.dependency 'PurchasesHybridCommon', '8.10.0-beta.8'
+
+  s.platform = :osx, '10.11'
+  s.pod_target_xcconfig = { 'DEFINES_MODULE' => 'YES' }
   s.swift_version = '5.0'
 end
