+++ purchases_ui_flutter/test/purchases_ui_flutter_test.dart
@@ -1,11 +1,11 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:purchases_flutter_ui/purchases_flutter_ui.dart';
+import 'package:purchases_ui_flutter/purchases_ui_flutter.dart';
 
 void main() {
   TestWidgetsFlutterBinding.ensureInitialized();
 
-  const channel = MethodChannel('purchases_flutter_ui');
+  const channel = MethodChannel('purchases_ui_flutter');
   final log = <MethodCall>[];
   dynamic response;
