+++ revenuecat_examples/purchase_tester/pubspec.yaml
@@ -20,8 +20,8 @@ dev_dependencies:
 
   purchases_flutter:
     path: ../../
-  purchases_flutter_ui:
-    path: ../../purchases_flutter_ui
+  purchases_ui_flutter:
+    path: ../../purchases_ui_flutter
 
   flutter_driver:
     sdk: flutter
